backend/requirements.txt
boto3>=1.34.0
psycopg2-binary>=2.9.9
sentence-transformers>=2.2.2
transformers>=4.35.0
torch>=2.0.0
pillow>=10.0.0
pgvector>=0.2.0
